refactor view page
